refactor: 타입 관련 폴더구조 정리
